switch to external globals entirely
