+++ caravel/globs.py
@@ -11,10 +11,23 @@ def init_globals():
     global log_path
     global act
     global compute_config
+    global logging_lvl
+    global summary_links
+    global dests
+    global reset_btn
+    global command
+    global currently_selected_package
     summarizer = None
     p = None
     selected_project = None
     log_path = None
     act = None
     compute_config = None
+    logging_lvl = None
+    summary_links = None
+    dests = None
+    reset_btn = None
+    command = None
+    currently_selected_package = None
+
