+++ caravel/jinja_templates/execute.html
@@ -22,6 +22,7 @@
 </script>
 {% endblock %}
 {% block content %}
-<div id="result" class="divider-vertical"><code>No result yet</code></div>
+<div id="result" class="divider-vertical"><b>Parsing the log file</b> <i class="fa fa-spinner fa-pulse fa-fw"></i></div>
+<div class="divider-vertical"></div>
 {% include 'back_referrer.html' %}
 {% endblock %}
